arrangement

+++ interface/server.h
@@ -3,5 +3,5 @@ typedef struct network_data{
 	uint8_t type;
 	uint64_t offset;
 	uint64_t len;
-	uint64_t seq;
+	uint32_t seq;
 }net_data_t;
